LayoutTests/fast/loader/javascript-url-in-embed.html

<script>
    if (window.layoutTestController)
        layoutTestController.dumpAsText();
</script>
<p>This tests that javascript: urls in an embed tag do not get executed.</p>
<embed src="javascript:alert('FAIL - javascript: url url was executed.');"></embed> 
